feat: Distance 테스트케이스 추가 및 구현

+++ src/main/java/racingcar/validation/DistanceValidator.java
@@ -8,7 +8,7 @@ public class DistanceValidator implements Validator{
 
     @Override
     public String validate(Object source) {
-        int distance = (int) source;
+        long distance = (long) source;
 
         if (distance >= 0) {
             return EMPTY_STR;
